Clean up src/client/logging/*
