+++ src/client/logging/index.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+
 'use strict';
 
 export {
@@ -12,5 +13,6 @@ export {
     logVerbose,
     logWarning,
     getPythonOutputChannelContent,
-    traceDecorators,
 } from './_global';
+
+export * as traceDecorators from './traceDecorators';
