+++ src/client/logging/util.ts
@@ -1,8 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+
 'use strict';
 
-export type Arguments = any[];
+import { Uri } from 'vscode';
+
+export type Arguments = unknown[];
 
 function valueToLogString(value: unknown, kind: string): string {
     if (value === undefined) {
@@ -12,8 +15,8 @@ function valueToLogString(value: unknown, kind: string): string {
         return 'null';
     }
     try {
-        if (value && (value as any).fsPath) {
-            return `<Uri:${(value as any).fsPath}>`;
+        if (value && (value as Uri).fsPath) {
+            return `<Uri:${(value as Uri).fsPath}>`;
         }
         return JSON.stringify(value);
     } catch {
